PYIC-1669: Rename journey engine lambda to be called process-journey-step
