--- lambdas/journeyengine/src/main/java/uk/gov/di/ipv/core/journeyengine/exceptions/JourneyEngineException.java
@@ -1,7 +1,0 @@
-package uk.gov.di.ipv.core.journeyengine.exceptions;
-
-public class JourneyEngineException extends Exception {
-    public JourneyEngineException(String errorMessage) {
-        super(errorMessage);
-    }
-}
