--- lambdas/journeyengine/src/main/java/uk/gov/di/ipv/core/journeyengine/statemachine/exceptions/UnknownEventException.java
@@ -1,23 +1,0 @@
-package uk.gov.di.ipv.core.journeyengine.statemachine.exceptions;
-
-import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
-
-@ExcludeFromGeneratedCoverageReport
-public class UnknownEventException extends Exception {
-
-    public UnknownEventException() {
-        super();
-    }
-
-    public UnknownEventException(String message) {
-        super(message);
-    }
-
-    public UnknownEventException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UnknownEventException(Throwable cause) {
-        super(cause);
-    }
-}
