--- lambdas/journeyengine/src/main/java/uk/gov/di/ipv/core/journeyengine/statemachine/responses/JourneyContext.java
@@ -1,12 +1,0 @@
-package uk.gov.di.ipv.core.journeyengine.statemachine.responses;
-
-import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
-
-@ExcludeFromGeneratedCoverageReport
-public class JourneyContext {
-    private JourneyContext() {}
-
-    public static JourneyContext emptyContext() {
-        return new JourneyContext();
-    }
-}
