+++ lambdas/process-journey-step/src/main/java/uk/gov/di/ipv/core/processjourneystep/exceptions/JourneyEngineException.java
@@ -1,0 +1,7 @@
+package uk.gov.di.ipv.core.processjourneystep.exceptions;
+
+public class JourneyEngineException extends Exception {
+    public JourneyEngineException(String errorMessage) {
+        super(errorMessage);
+    }
+}
