+++ lambdas/process-journey-step/src/main/java/uk/gov/di/ipv/core/processjourneystep/statemachine/exceptions/UnknownEventException.java
@@ -1,0 +1,23 @@
+package uk.gov.di.ipv.core.processjourneystep.statemachine.exceptions;
+
+import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
+
+@ExcludeFromGeneratedCoverageReport
+public class UnknownEventException extends Exception {
+
+    public UnknownEventException() {
+        super();
+    }
+
+    public UnknownEventException(String message) {
+        super(message);
+    }
+
+    public UnknownEventException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UnknownEventException(Throwable cause) {
+        super(cause);
+    }
+}
