+++ lambdas/process-journey-step/src/main/java/uk/gov/di/ipv/core/processjourneystep/statemachine/responses/JourneyContext.java
@@ -1,0 +1,12 @@
+package uk.gov.di.ipv.core.processjourneystep.statemachine.responses;
+
+import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
+
+@ExcludeFromGeneratedCoverageReport
+public class JourneyContext {
+    private JourneyContext() {}
+
+    public static JourneyContext emptyContext() {
+        return new JourneyContext();
+    }
+}
