+++ lambdas/process-journey-step/src/main/resources/IpvLambdaJsonLayout.json
