+++ lambdas/process-journey-step/src/main/resources/log4j2.yaml
