+++ lambdas/process-journey-step/src/main/resources/statemachine-config.yaml
