+++ settings.gradle
@@ -8,7 +8,7 @@ include "lib",
 		"lambdas:issuedcredentials",
 		"lambdas:sessionstart",
 		"lambdas:useridentity",
-		"lambdas:journeyengine",
+		"lambdas:process-journey-step",
 		"lambdas:credentialissuerstart",
 		"lambdas:credentialissuererror",
 		"lambdas:validatecricheck",
